#19: Fix class value

+++ Trabalho03/ClientApplication/src/Classes/AccommodationInterest.java
@@ -108,7 +108,7 @@ public class AccommodationInterest
      * 
      * @return  Number Of Guests
      */ 
-    public float getNumberOfGuests()
+    public int getNumberOfGuests()
     {
         return this.numberOfGuests;
     }
